@@ -1,6 +1,7 @@
 {
 	"name": "shopmagic-admin",
 	"version": "0.0.0",
+	"type": "module",
 	"scripts": {
 		"dev": "vite",
 		"build": "vite build",
@@ -10,7 +11,9 @@
 		"format": "prettier --write src/"
 	},
 	"dependencies": {
-		"pinia": "~2",
+		"color": "^5.0.0",
+		"dompurify": "^3.2.4",
+		"pinia": "~3",
 		"vue": "~3",
 		"vue-router": "~4"
 	},
@@ -18,40 +21,37 @@
 		"@jsonforms/core": "~3",
 		"@jsonforms/vue": "~3",
 		"@rushstack/eslint-patch": "~1",
-		"@tinymce/tinymce-vue": "~5",
+		"@tinymce/tinymce-vue": "~6",
 		"@tsconfig/node-lts": "^20",
-		"@types/color": "~3",
-		"@types/jsdom": "~21",
 		"@types/luxon": "~3",
 		"@types/node": "~20",
 		"@vicons/ionicons5": "~0",
 		"@vitejs/plugin-vue": "~5",
-		"@vue/eslint-config-prettier": "~9",
-		"@vue/eslint-config-typescript": "~12",
+		"@vue/eslint-config-prettier": "~10",
+		"@vue/eslint-config-typescript": "~14",
 		"@vue/test-utils": "~2",
-		"@vue/tsconfig": "~0.5",
-		"@vueuse/core": "~9",
-		"autoprefixer": "~10",
+		"@vue/tsconfig": "~0.7",
+		"@vueuse/core": "~13",
+		"autoprefixer": "^10.4.21",
 		"chart.js": "~4",
 		"chartjs-adapter-luxon": "~1",
-		"color": "~4",
 		"dayjs": "~1",
-		"eslint": "~8",
-		"eslint-config-async": "~2",
-		"eslint-plugin-vue": "~9",
-		"fuse.js": "~6",
-		"jsdom": "~23",
-		"lodash": "~4",
+		"eslint": "~9",
+		"eslint-config-async": "~3",
+		"eslint-plugin-vue": "~10",
+		"fuse.js": "~7",
+		"jsdom": "~26",
+		"lodash": "^4.17.21",
 		"luxon": "~3",
 		"naive-ui": "~2",
+		"postcss": "^8.5.3",
 		"prettier": "~3",
 		"swrv": "~1",
-		"tailwindcss": "~3",
-		"tinymce": "~6",
+		"tailwindcss": "^3.4.17",
 		"typescript": "~5",
-		"vite": "~5",
-		"vitest": "~1",
+		"vite": "~6",
+		"vitest": "~3",
 		"vue-chartjs": "~5",
-		"vue-tsc": "~1"
+		"vue-tsc": "~2"
 	}
 }
