@@ -1,7 +1,9 @@
-module.exports = {
+/** @type {import('tailwindcss').Config} */
+export default {
   content: ["./index.html", "./src/**/*.{vue,js,ts}"],
   theme: {
     extend: {},
   },
   plugins: [],
-};
+}
+
