@@ -145,7 +145,7 @@ async function deleteAutomation() {
       />
     </NLayoutHeader>
     <NLayout has-sider>
-      <NLayoutContent class="bg-gray-200">
+      <NLayoutContent class="bg-gray-200!">
         <EditGroup>
           <EventsPicker :events="events || []" :automation="automation" />
           <FilterEditor />
