@@ -1,5 +1,16 @@
 <script lang="ts" setup>
-import { NA, NButton, NP, NText, NIcon, NProgress, NSkeleton, NSpace, NSpin, NTable } from "naive-ui";
+import {
+  NA,
+  NButton,
+  NP,
+  NText,
+  NIcon,
+  NProgress,
+  NSkeleton,
+  NSpace,
+  NSpin,
+  NTable,
+} from "naive-ui";
 import { Checkmark, CloseOutline } from "@vicons/ionicons5";
 import { get } from "@/_utils";
 import { computed, ref, watchEffect } from "vue";
@@ -36,9 +47,10 @@ const BATCH_SIZE = 250;
 const page = ref(1);
 const maxPages = computed(() => Math.ceil(maxPossibleMatches.value / BATCH_SIZE));
 const dispatcherError = ref<object | null>(null);
-const automationName = computed(() => (
-  automation.value?.name ||
-    automation.value?.id && sprintf(__('Automation #%d'), automation.value?.id))
+const automationName = computed(
+  () =>
+    automation.value?.name ||
+    (automation.value?.id && sprintf(__("Automation #%d"), automation.value?.id)),
 );
 const automationActions = computed(() => {
   const rawActions = automation.value?.actions || [];
@@ -70,46 +82,43 @@ watchEffect(() => {
        * and collects errors. Check if we are able to use it.
        */
       async function determineApiMatchEndpoint() {
-        const { response } = await useWpFetch(`/automations/${props.id}/manual/match`).head()
-        if ( response.value.ok ) {
-          return `/automations/${props.id}/manual/match`
+        const { response } = await useWpFetch(`/automations/${props.id}/manual/match`).head();
+        if (response.value.ok) {
+          return `/automations/${props.id}/manual/match`;
         }
 
-        return `/automations/${props.id}/manual/matches`
+        return `/automations/${props.id}/manual/matches`;
       }
       const apiEndpoint = await determineApiMatchEndpoint();
       do {
         // TODO: abort controller when user leaves page
         await fetchPreviewRuns(apiEndpoint);
-      } while (page.value <= maxPages.value)
+      } while (page.value <= maxPages.value);
     })
-    .then(() => matchesLoading.value = false);
+    .then(() => (matchesLoading.value = false));
 });
 
 watchEffect(() => {
-  if ( matchingItems.value.length !== 0 ) {
-    matchesLoading.value = false
+  if (matchingItems.value.length !== 0) {
+    matchesLoading.value = false;
   }
-})
+});
 
 async function fetchPreviewRuns(apiMatchEndpoint: string) {
   const CONCURRENT_REQS = 5;
   const batchSize = Math.min(CONCURRENT_REQS, maxPages.value - page.value + 1);
-  const promises = Array.from({length: batchSize}, (_, i) => {
+  const promises = Array.from({ length: batchSize }, (_, i) => {
     return get(
-      appendSearchParams(
-        apiMatchEndpoint,
-        {
-          page_size: BATCH_SIZE,
-          page: page.value + i,
-        }
-      )
-    )
+      appendSearchParams(apiMatchEndpoint, {
+        page_size: BATCH_SIZE,
+        page: page.value + i,
+      }),
+    );
   });
   const result = await Promise.all(promises);
   page.value += batchSize;
-  if( apiMatchEndpoint.endsWith(`automations/${props.id}/manual/match`) ) {
-    for (const {data: matches, errors: matchErrors} of result) {
+  if (apiMatchEndpoint.endsWith(`automations/${props.id}/manual/match`)) {
+    for (const { data: matches, errors: matchErrors } of result) {
       matchingItems.value.push(...matches);
       matchingErrors.value.push(...matchErrors);
       processed.value++;
@@ -159,7 +168,7 @@ async function dispatchAutomations() {
   <EditGroup>
     <ShadyCard>
       <template #header>
-        {{ __("Manual trigger sources for:", 'shopmagic-for-woocommerce') + " " }}
+        {{ __("Manual trigger sources for:", "shopmagic-for-woocommerce") + " " }}
         <span v-if="automationName">
           {{ automationName }}
         </span>
@@ -169,10 +178,19 @@ async function dispatchAutomations() {
         <ul class="h-[300px] overflow-y-scroll shadow-inner p-1">
           <li v-if="!matchesLoading && matchingItems.length === 0" class="p-3">
             <NText type="warning">
-              {{ __('There are no data sources matching current automation settings. Make sure your settings are correct and rerun the process.', 'shopmagic-for-woocommerce') }}
+              {{
+                __(
+                  "There are no data sources matching current automation settings. Make sure your settings are correct and rerun the process.",
+                  "shopmagic-for-woocommerce",
+                )
+              }}
             </NText>
           </li>
-          <li v-for="item in matchingItems" :key="item.id" class="p-1 rounded [&:nth-child(2n)]:bg-gray-50">
+          <li
+            v-for="item in matchingItems"
+            :key="item.id"
+            class="p-1 rounded-sm nth-[2n]:bg-gray-50"
+          >
             <NA v-if="item.link" :href="item.link" target="_blank">
               {{ item.name }} #{{ item.id }}
             </NA>
@@ -181,10 +199,28 @@ async function dispatchAutomations() {
         </ul>
       </NSpin>
       <template #action>
-        <NP>{{ sprintf( __( "Total items: %d", 'shopmagic-for-woocommerce' ), matchingItems.length ) }}</NP>
+        <NP>{{
+          sprintf(__("Total items: %d", "shopmagic-for-woocommerce"), matchingItems.length)
+        }}</NP>
         <NP type="error" v-if="matchingErrors.length > 0">
           <NText type="error">
-            {{ sprintf( _n("There was %d error, when trying to match sources applicable for automation.", "There were %d errors, when trying to match sources applicable for automation.", matchingErrors.length, 'shopmagic-for-woocommerce'), matchingErrors.length ) }} {{ __("You still can execute the automation with all errorneous sources skipped. Otherwise, change your automation settings or check ShopMagic's error logs.", 'shopmagic-for-woocommerce') }}
+            {{
+              sprintf(
+                _n(
+                  "There was %d error, when trying to match sources applicable for automation.",
+                  "There were %d errors, when trying to match sources applicable for automation.",
+                  matchingErrors.length,
+                  "shopmagic-for-woocommerce",
+                ),
+                matchingErrors.length,
+              )
+            }}
+            {{
+              __(
+                "You still can execute the automation with all errorneous sources skipped. Otherwise, change your automation settings or check ShopMagic's error logs.",
+                "shopmagic-for-woocommerce",
+              )
+            }}
           </NText>
         </NP>
         <NProgress
@@ -223,7 +259,9 @@ async function dispatchAutomations() {
     <NSpace>
       <NButton type="primary" @click="dispatchAutomations">
         <NSpin :show="queueDispatched === STATES.LOADING" :size="12" stroke="#eaeaea">
-          <span v-if="queueDispatched === STATES.NOT_INITIALIZED">{{ __("Run actions", 'shopmagic-for-woocommerce' ) }}</span>
+          <span v-if="queueDispatched === STATES.NOT_INITIALIZED">{{
+            __("Run actions", "shopmagic-for-woocommerce")
+          }}</span>
           <span v-else-if="queueDispatched === STATES.SUCCESS">
             <NIcon><Checkmark /> </NIcon>
           </span>
