@@ -1,3 +1,31 @@
+<script lang="ts" setup>
+import { NCheckbox, NCheckboxGroup } from "naive-ui";
+import {
+  type ControlElement,
+  mapDispatchToControlProps,
+  mapStateToControlProps,
+} from "@jsonforms/core";
+import { computed } from "vue";
+import { rendererProps, useJsonFormsOneOfEnumControl, useControl } from "@jsonforms/vue";
+import { useVanillaControl } from "../util";
+import FieldWrapper from "./FieldWrapper.vue";
+
+const props = defineProps(rendererProps<ControlElement>());
+const { control, controlWrapper, onChange } = useVanillaControl(
+  useControl(props, mapStateToControlProps, mapDispatchToControlProps),
+);
+
+const options = computed(() => {
+  const items = control.value.schema.items?.oneOf || control.value.schema.items?.anyOf || [];
+
+  return items
+    .filter((s) => s.title)
+    .map((s) => ({
+      label: s.title,
+      value: s.const,
+    }));
+});
+</script>
 <template>
   <FieldWrapper v-bind="controlWrapper">
     <NCheckboxGroup
@@ -6,7 +34,7 @@
       @update:value="onChange"
     >
       <NCheckbox
-        v-for="(checkbox, i) in control.options"
+        v-for="(checkbox, i) in options"
         :key="i"
         :label="checkbox.label"
         :value="checkbox.value"
@@ -15,27 +43,3 @@
     </NCheckboxGroup>
   </FieldWrapper>
 </template>
-
-<script lang="ts">
-import { NCheckbox, NCheckboxGroup } from "naive-ui";
-import type { ControlElement } from "@jsonforms/core";
-import { defineComponent } from "vue";
-import { rendererProps, type RendererProps, useJsonFormsOneOfEnumControl } from "@jsonforms/vue";
-import { useVanillaControl } from "../util";
-import FieldWrapper from "./FieldWrapper.vue";
-
-export default defineComponent({
-  name: "MultipleChoiceCheckboxControlRenderer",
-  components: {
-    FieldWrapper,
-    NCheckbox,
-    NCheckboxGroup,
-  },
-  props: {
-    ...rendererProps<ControlElement>(),
-  },
-  setup(props: RendererProps<ControlElement>) {
-    return useVanillaControl(useJsonFormsOneOfEnumControl(props), (target) => target);
-  },
-});
-</script>
