@@ -1,13 +1,10 @@
-import StringControlRenderer from "@/components/FormRenderers/controls/StringControlRenderer.vue";
 import {
   and,
   isBooleanControl,
-  isDateControl,
   isDateTimeControl,
-  isEnumControl,
-  isIntegerControl,
   isNumberControl,
-  isOneOfEnumControl,
+  isOneOfControl,
+  isAnyOfControl,
   isScoped,
   isStringControl,
   isTimeControl,
@@ -15,31 +12,30 @@ import {
   or,
   rankWith,
   resolveSchema,
+  schemaTypeIs,
   uiTypeIs,
 } from "@jsonforms/core";
-import GoogleWorksheetsControlRenderer from "@/components/FormRenderers/controls/GoogleWorksheetsControlRenderer.vue";
-import ActionControlRenderer from "@/components/FormRenderers/controls/ActionControlRenderer.vue";
-import MultipleChoiceCheckboxControlRenderer from "@/components/FormRenderers/controls/MultipleChoiceCheckboxControlRenderer.vue";
-import EnumOneOfControlRenderer from "@/components/FormRenderers/controls/EnumOneOfControlRenderer.vue";
-import TimeControlRenderer from "@/components/FormRenderers/controls/TimeControlRenderer.vue";
-import MediaPickerControlRenderer from "@/components/FormRenderers/controls/MediaPickerControlRenderer.vue";
-import NumberControlRenderer from "@/components/FormRenderers/controls/NumberControlRenderer.vue";
-import DateControlRenderer from "@/components/FormRenderers/controls/DateControlRenderer.vue";
-import TextControlRenderer from "@/components/FormRenderers/controls/TextControlRenderer.vue";
-import EditorControlRenderer from "@/components/FormRenderers/controls/EditorControlRenderer.vue";
-import BooleanControlRenderer from "@/components/FormRenderers/controls/BooleanControlRenderer.vue";
-import DateTimeControlRenderer from "@/components/FormRenderers/controls/DateTimeControlRenderer.vue";
-import EnumControlRenderer from "@/components/FormRenderers/controls/EnumControlRenderer.vue";
-import IntegerControlRenderer from "@/components/FormRenderers/controls/IntegerControlRenderer.vue";
+import GenericControlRenderer from "./GenericControlRenderer.vue";
+import GoogleWorksheetsControlRenderer from "./GoogleWorksheetsControlRenderer.vue";
+import ActionControlRenderer from "./ActionControlRenderer.vue";
+import MultipleChoiceCheckboxControlRenderer from "./MultipleChoiceCheckboxControlRenderer.vue";
+import SelectControlRenderer from "./SelectControlRenderer.vue";
+import TimeControlRenderer from "./TimeControlRenderer.vue";
+import MediaPickerControlRenderer from "./MediaPickerControlRenderer.vue";
+import NumberControlRenderer from "./NumberControlRenderer.vue";
+import TextControlRenderer from "./TextControlRenderer.vue";
+import EditorControlRenderer from "./EditorControlRenderer.vue";
+import CheckboxControlRenderer from "./CheckboxControlRenderer.vue";
+import DateTimeControlRenderer from "./DateTimeControlRenderer.vue";
 import ManualActionControlRenderer from "./ManualActionControlRenderer.vue";
 import PluginModuleRenderer from "./PluginModuleRenderer.vue";
 import ButtonControlRenderer from "./ButtonControlRenderer.vue";
 import RawHtmlControlRenderer from "./RawHtmlControlRenderer.vue";
-import TextAreaControlRenderer from "@/components/FormRenderers/controls/TextAreaControlRenderer.vue";
+import ProductArrayRenderer from "./ProductArrayRenderer.vue";
 
 export const controlRenderers: JsonFormsRendererRegistryEntry[] = [
   {
-    renderer: StringControlRenderer,
+    renderer: GenericControlRenderer,
     tester: rankWith(1, isStringControl),
   },
   {
@@ -47,20 +43,8 @@ export const controlRenderers: JsonFormsRendererRegistryEntry[] = [
     tester: rankWith(1, isNumberControl),
   },
   {
-    renderer: IntegerControlRenderer,
-    tester: rankWith(1, isIntegerControl),
-  },
-  {
-    renderer: EnumControlRenderer,
-    tester: rankWith(2, isEnumControl),
-  },
-  {
-    renderer: EnumOneOfControlRenderer,
-    tester: rankWith(2, isOneOfEnumControl),
-  },
-  {
-    renderer: DateControlRenderer,
-    tester: rankWith(2, isDateControl),
+    renderer: SelectControlRenderer,
+    tester: rankWith(2, or(isOneOfControl, isAnyOfControl, schemaTypeIs("array"))),
   },
   {
     renderer: DateTimeControlRenderer,
@@ -71,15 +55,7 @@ export const controlRenderers: JsonFormsRendererRegistryEntry[] = [
     tester: rankWith(2, isTimeControl),
   },
   {
-    renderer: PluginModuleRenderer,
-    tester: rankWith(3, (uischema, schema, context) => {
-      if (!isScoped(uischema)) return false;
-      const elementSchema = resolveSchema(schema, uischema.scope, context.rootSchema);
-      return elementSchema?.format === "plugin-module";
-    }),
-  },
-  {
-    renderer: BooleanControlRenderer,
+    renderer: CheckboxControlRenderer,
     tester: rankWith(
       3,
       or(isBooleanControl, (uischema, schema, context) => {
@@ -93,12 +69,20 @@ export const controlRenderers: JsonFormsRendererRegistryEntry[] = [
     renderer: MultipleChoiceCheckboxControlRenderer,
     tester: rankWith(
       3,
-      and(isOneOfEnumControl, (uischema, schema, context) => {
+      and(schemaTypeIs("array"), (uischema, schema, context) => {
         if (!isScoped(uischema)) return false;
         const elementSchema = resolveSchema(schema, uischema.scope, context.rootSchema);
-        return (elementSchema.uniqueItems && elementSchema?.format !== "select") || false;
+        return elementSchema?.format === "checkbox";
       }),
     ),
+  },
+  {
+    renderer: PluginModuleRenderer,
+    tester: rankWith(3, (uischema, schema, context) => {
+      if (!isScoped(uischema)) return false;
+      const elementSchema = resolveSchema(schema, uischema.scope, context.rootSchema);
+      return elementSchema?.format === "plugin-module";
+    }),
   },
   {
     renderer: MediaPickerControlRenderer,
@@ -108,20 +92,6 @@ export const controlRenderers: JsonFormsRendererRegistryEntry[] = [
         if (!isScoped(uischema)) return false;
         const elementSchema = resolveSchema(schema, uischema.scope, context.rootSchema);
         return elementSchema?.format === "file";
-      }),
-    ),
-  },
-  {
-    renderer: TextAreaControlRenderer,
-    tester: rankWith(
-      2,
-      and(isStringControl, (uischema, schema, context) => {
-        if (!isScoped(uischema)) return false;
-        const elementSchema = resolveSchema(schema, uischema.scope, context.rootSchema);
-        return (
-          (elementSchema?.format === "textarea" && elementSchema?.presentation?.type === "plain") ||
-          false
-        );
       }),
     ),
   },
@@ -185,7 +155,7 @@ export const controlRenderers: JsonFormsRendererRegistryEntry[] = [
     tester: rankWith(5, (uischema, schema, context) => {
       if (!isScoped(uischema)) return false;
       const elementSchema = resolveSchema(schema, uischema.scope, context.rootSchema);
-      return elementSchema?.format === "advertisement" || elementSchema.type === "null";
+      return elementSchema?.format === "advertisement";
     }),
   },
   {
@@ -195,5 +165,16 @@ export const controlRenderers: JsonFormsRendererRegistryEntry[] = [
       const elementSchema = resolveSchema(schema, uischema.scope, context.rootSchema);
       return elementSchema?.format === "button" && elementSchema.type === "null";
     }),
+  },
+  {
+    renderer: ProductArrayRenderer,
+    tester: rankWith(
+      10,
+      and(schemaTypeIs("array"), (uischema, schema, context) => {
+        if (!isScoped(uischema)) return false;
+        const elementSchema = resolveSchema(schema, uischema.scope, context.rootSchema);
+        return elementSchema?.presentation.type === "products";
+      }),
+    ),
   },
 ];
