@@ -5,22 +5,24 @@ import vue from "@vitejs/plugin-vue";
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [vue()],
+  plugins: [
+    vue(),
+  ],
   resolve: {
     alias: {
       "@": fileURLToPath(new URL("./src", import.meta.url)),
     },
   },
   base: "",
-  optimizeDeps: {
-    disabled: "build",
-  },
   esbuild: {
     /**
      * WordPress doesn't parse esbuild minification correctly.
      * Due to that don't optimize identifiers for better i18n handling.
      */
     minifyIdentifiers: false,
+  },
+  server: {
+    cors: true,
   },
   build: {
     manifest: true,
